+++ Tests/Remove-CHttpsCertificateBinding.Tests.ps1
@@ -19,52 +19,155 @@ BeforeAll {
 
     & (Join-Path -Path $PSScriptRoot -ChildPath 'Initialize-Test.ps1' -Resolve)
 
-    $cert = $null
     $ipAddress = '1.2.3.4'
-    $ipV6Address = '::1234'
-    $port = '8483'
-    $ipPort = '{0}:{1}' -f $ipAddress,$port
-    $ipv6Port = '[{0}]:{1}' -f $ipV6Address,$port
+    $script:ipNum = 1
     $appID = '454f19a6-3ea8-434c-874f-3a860778e4af'
-    $ipV6AppID = 'b01fa31e-d255-48df-983e-c5c6dd0ccd03'
+
+    $certPath = Join-Path -Path $PSScriptRoot -ChildPath 'CarbonTestCertificate.pfx' -Resolve
+    $script:cert = Install-CCertificate -Path $certPath -StoreLocation LocalMachine -StoreName My -PassThru
+
+    $lastBinding = Get-CHttpsCertificateBinding | Sort-Object -Property 'Port' | Sort-Object | Select-Object -Last 1
+    $script:port = 44444
+    if ($lastBinding)
+    {
+        $script:port = $lastBinding.Port + 1
+    }
+    $script:appId = '9236e152-2b3e-4674-a410-58f2d218c66c'
+
+    function New-IPAddress
+    {
+        param(
+            [switch] $V6
+        )
+
+        [ipaddress] $ipAddress = "$($script:ipNum).$($script:ipNum).$($script:ipNum).$($script:ipNum)"
+        if ($V6)
+        {
+            $ipAddress = "::$($script:ipNum)"
+        }
+        $script:ipNum += 1
+        return $ipAddress
+    }
+
+    function GivenBinding
+    {
+        param(
+            [ipaddress] $IPAddress = (New-IPAddress),
+
+            [UInt64] $Port = $script:port
+        )
+
+        Set-CHttpsCertificateBinding -IPAddress $IPAddress `
+                                   -Port $Port `
+                                   -Thumbprint $script:cert.Thumbprint `
+                                   -ApplicationID $script:appID
+    }
+
+    function ThenNoErrors
+    {
+        $Global:Error | Should -BeNullOrEmpty
+    }
+}
+
+AfterAll {
+    Remove-CHttpsCertificateBinding -ApplicationID $script:appID -ErrorAction Ignore
+    Remove-CHttpsCertificateBinding -Thumbprint $script:cert.Thumbprint -ErrorAction Ignore
+    Uninstall-CCertificate -Certificate $script:cert -StoreLocation LocalMachine -StoreName My
 }
 
 Describe 'Remove-CHttpsCertificateBinding' {
     BeforeEach {
-        $cert = Install-CCertificate -Path (Join-Path -Path $PSScriptRoot -ChildPath 'CarbonTestCertificate.cer' -Resolve) `
-                                     -StoreLocation LocalMachine `
-                                     -StoreName My `
-                                     -PassThru
-        netsh http add sslcert ipport=$ipPort "certhash=$($cert.Thumbprint)" "appid={$appID}"
-        netsh http add sslcert ipport=$ipV6Port "certhash=$($cert.Thumbprint)" "appid={$ipV6AppID}"
+        $Global:Error.Clear()
     }
 
     AfterEach {
-        netsh http delete sslcert ipport=$ipPort
-        netsh http delete sslcerrt ipport=$ipV6Port
+        Remove-CHttpsCertificateBinding -ApplicationID $script:appID -ErrorAction Ignore
+        Remove-CHttpsCertificateBinding -Thumbprint $script:cert.Thumbprint -ErrorAction Ignore
 
-        Uninstall-CCertificate -Certificate $cert -StoreLocation LocalMachine -StoreName My
+        $script:port += 1
     }
 
     It 'should remove non existent binding' {
-        $bindings = @( Get-CHttpsCertificateBinding )
-        Remove-CHttpsCertificateBinding -IPAddress '1.2.3.4' -Port '8332'
-        $newBindings = @( Get-CHttpsCertificateBinding )
-        $newBindings.Length | Should -Be $bindings.Length
+        $currentBindings = Get-CHttpsCertificateBinding
+        Remove-CHttpsCertificateBinding -IPAddress (New-IPAddress) -ErrorAction SilentlyContinue
+        Remove-CHttpsCertificateBinding -IPAddress (New-IPAddress -V6) -ErrorAction SilentlyContinue
+        Remove-CHttpsCertificateBinding -Port $script:port -ErrorAction SilentlyContinue
+        Remove-CHttpsCertificateBinding -Thumbprint 'deadbee' -ErrorAction SilentlyContinue
+        Remove-CHttpsCertificateBinding -ApplicationID ([Guid]::NewGuid()) -ErrorAction SilentlyContinue
+        $Global:Error | Should -Match 'because it does not exist'
+        Get-CHttpsCertificateBinding | Should -HaveCount $currentBindings.Count
+
+        $Global:Error.Clear()
+        Remove-CHttpsCertificateBinding -IPAddress (New-IPAddress) -ErrorAction Ignore
+        $Global:Error | Should -BeNullOrEmpty
     }
 
-    It 'should not remove certificate what if' {
-        Remove-CHttpsCertificateBinding -IPAddress $ipAddress -Port $port -WhatIf
-        (Test-CHttpsCertificateBinding -IPAddress $ipAddress -Port $port) | Should -BeTrue
+    It 'should not remove certificate' {
+        GivenBinding
+        Remove-CHttpsCertificateBinding -ApplicationID $script:appID -WhatIf
+        ThenNoErrors
+        Test-CHttpsCertificateBinding -ApplicationID $script:appID | Should -BeTrue
     }
 
-    It 'should remove binding' {
-        Remove-CHttpsCertificateBinding -IPAddress $ipAddress -Port $port
-        (Test-CHttpsCertificateBinding -IPAddress $ipAddress -Port $port) | Should -BeFalse
+    It 'should remove binding by IP v4 address' {
+        $ipAddress = New-IPAddress
+        GivenBinding -IPAddress $ipAddress
+        Remove-CHttpsCertificateBinding -IPAddress $ipAddress
+        ThenNoErrors
+        Test-CHttpsCertificateBinding -IPAddress $ipAddress | Should -BeFalse
     }
 
-    It 'should remove i pv6 binding' {
-        Remove-CHttpsCertificateBinding -IPAddress $ipV6Address -Port $port
-        (Test-CHttpsCertificateBinding -IPAddress $ipV6Address -Port $port) | Should -BeFalse
+    It 'should remove binding by IP v6 address' {
+        $ipAddress = New-IPAddress -V6
+        GivenBinding -IPAddress $ipAddress
+        Remove-CHttpsCertificateBinding -IPAddress $ipAddress
+        ThenNoErrors
+        Test-CHttpsCertificateBinding -IPAddress $ipAddress | Should -BeFalse
+    }
+
+    It 'should remove binding by certificate thumbprint' {
+        GivenBinding
+        Remove-CHttpsCertificateBinding -Thumbprint $script:cert.Thumbprint
+        ThenNoErrors
+        Test-CHttpsCertificateBinding -Thumbprint $script:cert.Thumbprint | Should -BeFalse
+    }
+
+    It 'should remove binding by application ID' {
+        GivenBinding
+        Remove-CHttpsCertificateBinding -ApplicationID $script:appId
+        ThenNoErrors
+        Test-CHttpsCertificateBinding -ApplicationID $script:appId | Should -BeFalse
+    }
+
+    It 'should remove binding using multiple criteria' {
+        $ip1 = New-IPAddress
+        GivenBinding -IPAddress $ip1
+        $ip2 = New-IPAddress
+        GivenBinding -IPAddress $ip2
+        Remove-CHttpsCertificateBinding -IPAddress $ip1 -Port $script:port
+        ThenNoErrors
+        Test-CHttpsCertificateBinding -IPAddress $ip1 | Should -BeFalse
+        Test-CHttpsCertificateBinding -IPAddress $ip2 | Should -BeTrue
+    }
+
+    It 'should accept pipeline input' {
+        GivenBinding
+        GivenBinding
+        Get-CHttpsCertificateBinding -ApplicationID $script:appID | Should -HaveCount 2
+        Get-CHttpsCertificateBinding -ApplicationID $script:appID | Remove-CHttpsCertificateBinding
+        Get-CHttpsCertificateBinding -ApplicationID $script:appID -ErrorAction Ignore | Should -BeNullOrEmpty
+    }
+
+    It 'should remove all bindings' {
+        GivenBinding
+        $bindings = Get-CHttpsCertificateBinding
+        Mock -CommandName 'Invoke-Netsh' -ModuleName 'Carbon.Windows.HttpServer'
+        Remove-CHttpsCertificateBinding -Force
+        Assert-MockCalled -CommandName 'Invoke-Netsh' -ModuleName 'Carbon.Windows.HttpServer' -ParameterFilter {
+            $ArgumentList[0] | Should -Be 'http'
+            $ArgumentList[1] | Should -Be 'delete'
+            $true
+        } -Times $bindings.Count -Exactly
+        ThenNoErrors
     }
 }